+++ tests/test_central_pydevccu.py
@@ -28,8 +28,8 @@ async def test_central_mini(central_unit_mini) -> None:
     assert central_unit_mini.model == "PyDevCCU"
     assert central_unit_mini.get_client(const.INTERFACE_ID).model == "PyDevCCU"
     assert central_unit_mini.primary_client.model == "PyDevCCU"
-    assert len(central_unit_mini._devices) == 1
-    assert len(central_unit_mini.get_entities(exclude_no_create=False)) == 34
+    assert len(central_unit_mini._devices) == 2
+    assert len(central_unit_mini.get_entities(exclude_no_create=False)) == 63
 
 
 @pytest.mark.asyncio()