+++ changelog.md
@@ -1,5 +1,6 @@
 # Version 2024.10.7 (2024-10-12)
 
+- Add OPTIMUM_START_STOP and TEMPERATURE_OFFSET to climate
 - Improve profile validation
 - Use regex to identify schedule profiles